requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
openai
httpx
aiofiles
PyJWT
passlib
pydantic[email]
bcrypt==4.0.1